==> scripts/05a_setup_gridded_params_weddell.ipynb <==
# Processing inputs for Weddell Sea gridded DBPM run at $0.25^{\circ}$  
**Author**: Denisse Fierro Arcos  
**Date**: 2025-04-28  
  
We encountered issues will running the DBPM at $0.25^{\circ}$ in the Weddell Sea. Both fish groups represented in the model, benthic detritivores and pelagic predators, crashed in the 1880s. We identified this was due to numerical instabilities.   
  
Since the model ran for the Weddell Sea at $1^{\circ}$, we used this coarser resolution to find the largest temporal change in inputs that the model was able to handle. We applied at multiplier of 0.90 to the largest temporal change because numerical instabilities remained an issue.  
  
All functions used here are included in the `useful_functions.py` scripts.

## Loading relevant libraries

import os
os.chdir('/g/data/vf71/la6889/dbpm_southern_ocean/scripts/')
import xarray as xr
from glob import glob
import numpy as np
import useful_functions as uf

## Defining basic variables to apply input correction

base_folder = '/g/data/vf71/la6889/dbpm_inputs/weddell/gridded'
variables = ['lphy', 'sphy']
experiments = ['ctrlclim', 'obsclim', 'spinup', 'stable-spin']

## Applying corrections

for var in variables:
    lmad = uf.lowest_maximum_absolute_diff(base_folder, '1deg', var, prop = 0.9)
    for exp in experiments:
        uf.mask_absolute_diff(base_folder, '025deg', exp, var, threshold = lmad)
        uf.exportRatio_intercept_slope(base_folder, '025deg', exp)
